pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Rapidbay"
version = "1.0.0"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "aiohttp==3.9.5",
    "bencodepy==0.9.5",
    "diskcache==5.6.3",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "iso-639==0.4.5",
    "lxml==5.4.0",
    "parse-torrent-name==1.1.1",
    "pymediainfo==4.2.1",
    "python-dateutil==2.9.0.post0",
    "python-opensubtitles @ git+https://github.com/agonzalezro/python-opensubtitles@0143db67df5b80155c258887ca36f62a92455d4e",
    "requests==2.31.0",
    "urllib3==1.26.16",
    "libtorrent==2.0.11",
    "types-libtorrent==2.0.11.20250816",
]

[project.optional-dependencies]
dev = [
    "basedpyright==1.31.3",
    "types-click==7.1.8",
    "types-html5lib==1.1.11.20241018",
    "types-Jinja2==2.11.9",
    "types-libtorrent==2.0.11.20250816",
    "types-lxml==2025.8.25",
    "types-MarkupSafe==1.1.10",
    "types-python-dateutil==2.9.0.20241206",
    "types-requests==2.28.11.17",
    "types-setuptools==75.8.0.20250110",
    "types-urllib3==1.26.25.14",
    "ruff==0.12.11",
    "pytest==8.3.4",
]

[tool.basedpyright]
# Python version - matching project Python version  
pythonVersion = "3.9"
pythonPlatform = "All"

# Import discovery
exclude = ["**/frontend/**"]

# Type checking mode
typeCheckingMode = "standard"

# Warnings configuration
reportUnnecessaryCast = true  # warn_redundant_casts = True
reportUnnecessaryTypeIgnoreComment = true  # warn_unused_ignores = True
reportUnusedImport = true  # warn_unused_configs = True
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true

# Error strictness configuration
reportUntypedFunctionDecorator = "none"  # disallow_untyped_decorators = False
reportUntypedClassDecorator = "none"  # disallow_untyped_decorators = False
reportUntypedBaseClass = "none"
reportUnknownParameterType = "error"  # disallow_untyped_defs = True
reportUnknownArgumentType = "none"  # disallow_untyped_calls = False
reportUnknownLambdaType = "error"  # disallow_untyped_defs = True
reportUnknownVariableType = "error"  # check_untyped_defs = True
reportUnknownMemberType = "none"
reportMissingParameterType = "error"  # disallow_incomplete_defs = True
reportMissingTypeArgument = "warning"

# None/Optional handling configuration
reportImplicitStringConcatenation = "warning"
reportCallInDefaultInitializer = "none"
reportUnnecessaryIsInstance = true  # strict_equality = True
reportUnnecessaryComparison = true  # strict_equality = True
reportImplicitOverride = "warning"
reportShadowedImports = "warning"

# Additional strictness configuration
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
reportOverlappingOverload = true
reportPossiblyUnboundVariable = true  # warn_unreachable = True
reportConstantRedefinition = "error"
reportInvalidStringEscapeSequence = "error"
reportMissingTypeStubs = false
reportPrivateUsage = "warning"
reportTypeCommentUsage = "warning"
reportInvalidTypeVarUse = "error"
reportAssertAlwaysTrue = "warning"
reportSelfClsParameterName = "error"
reportUnusedExpression = "warning"
reportWildcardImportFromLibrary = "warning"
reportGeneralTypeIssues = "error"

# Allow old-style typing imports (Dict, List, Optional, Union)
reportDeprecated = false

# Handle missing imports - for libraries without stubs
reportMissingImports = "none"  # Matches ignore_missing_imports = True for external libs

# Execution environment
executionEnvironments = [
    { root = "app", pythonVersion = "3.9", extraPaths = ["."] }
]

# Additional configuration for handling imports
stubPath = "stubs"
include = ["app", "stubs"]
useLibraryCodeForTypes = true

[tool.ruff]
# Ruff configuration
line-length = 120
target-version = "py39"

[tool.ruff.lint]
# Enable pycodestyle, pyflakes, and more
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # line too long (handled by line-length)
    "E203",  # whitespace before ':'
    "UP006", # use list instead of List (keep old-style for consistency)
    "UP035", # import from collections.abc (keep typing imports for consistency)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Exclude directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".tox",
    "dist",
    "build",
    "*.egg",
    "app/frontend",
]

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files
"app/frontend/**" = ["F821"]  # Ignore undefined names in Kodi frontend

[tool.ruff.lint.isort]
known-first-party = ["app"]

# Format settings
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]

[tool.setuptools]
packages = ["app"]
